// frontend/src/api/generated/models/pageResponsePublicationResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * ComeSebebes API
 * API RESTful da rede social de comida ComeSebebes.
 * OpenAPI spec version: v1
 */
import type { PublicationResponse } from './publicationResponse';

export interface PageResponsePublicationResponse {
  content: PublicationResponse[];
  page: number;
  size: number;
  totalElements: number;
  totalPages: number;
  first: boolean;
  last: boolean;
}